+++ common/src/main/java/net/mehvahdjukaar/every_compat/modules/quark/CompatChestBlock.java
@@ -27,7 +27,7 @@ public class CompatChestBlock extends VariantChestBlock {
     }
 
     public String getTextureFolder() {
-        return "model/chest/everycompat";
+        return "quark_variant_chests/everycompat";
     }
 
     public String getTexturePath(){