+++ common/src/main/java/net/mehvahdjukaar/every_compat/modules/quark/QuarkModule.java
@@ -1,6 +1,8 @@
 package net.mehvahdjukaar.every_compat.modules.quark;
 
 import com.mojang.datafixers.util.Pair;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.mehvahdjukaar.every_compat.ECPlatformStuff;
 import net.mehvahdjukaar.every_compat.EveryCompat;
 import net.mehvahdjukaar.every_compat.api.SimpleEntrySet;
@@ -33,10 +35,12 @@ import net.minecraft.world.level.block.ComposterBlock;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.entity.ChestBlockEntity;
 import net.minecraft.world.level.block.state.BlockBehaviour;
+import org.violetmoon.quark.base.QuarkClient;
 import org.violetmoon.quark.content.building.block.*;
 import org.violetmoon.quark.content.building.client.render.be.VariantChestRenderer;
 import org.violetmoon.quark.content.building.module.*;
 import org.violetmoon.zeta.block.ZetaBlock;
+import org.violetmoon.zeta.client.SimpleWithoutLevelRenderer;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -222,7 +226,7 @@ public class QuarkModule extends SimpleModule {
                         () -> LeavesTypeRegistry.OAK_TYPE,
                         (w) -> {
                             if (w.getWoodType() == null) return null;
-                            return new HedgeBlock(null, Blocks.OAK_FENCE, w.leaves);
+                            return new HedgeBlock("",null, Blocks.OAK_FENCE, w.leaves);
                         })
                 .addModelTransform(m -> m.replaceWithTextureFromChild("minecraft:block/oak_leaves",
                         "leaves", s -> !s.contains("/snow") && !s.contains("_snow")))
@@ -291,6 +295,19 @@ public class QuarkModule extends SimpleModule {
         TRAPPED_CHEST_TILE = trappedChests.getTile(ChestBlockEntity.class);
     }
 
+    @Environment(EnvType.CLIENT)
+    @Override
+    public void onClientSetup() {
+        super.onClientSetup();
+
+        for(var b : chests.blocks.values())
+            QuarkClient.ZETA_CLIENT.setBlockEntityWithoutLevelRenderer(b.asItem(), new SimpleWithoutLevelRenderer(CHEST_TILE, b.defaultBlockState()));
+        for(var b : trappedChests.blocks.values())
+            QuarkClient.ZETA_CLIENT.setBlockEntityWithoutLevelRenderer(b.asItem(), new SimpleWithoutLevelRenderer(TRAPPED_CHEST_TILE, b.defaultBlockState()));
+
+    }
+
+    @Environment(EnvType.CLIENT)
     @Override
     public void registerBlockEntityRenderers(ClientHelper.BlockEntityRendererEvent event) {
         super.registerBlockEntityRenderers(event);
@@ -321,13 +338,13 @@ public class QuarkModule extends SimpleModule {
     public void addDynamicClientResources(ClientDynamicResourcesHandler handler, ResourceManager manager) {
         super.addDynamicClientResources(handler, manager);
 
-        try (TextureImage normal = TextureImage.open(manager, modRes("model/chest/oak/normal"));
+        try (TextureImage normal = TextureImage.open(manager, modRes("quark_variant_chests/oak/normal"));
              TextureImage normal_m = TextureImage.open(manager, EveryCompat.res("model/oak_chest_normal_m"));
              TextureImage normal_o = TextureImage.open(manager, EveryCompat.res("model/oak_chest_normal_o"));
-             TextureImage left = TextureImage.open(manager, modRes("model/chest/oak/left"));
+             TextureImage left = TextureImage.open(manager, modRes("quark_variant_chests/oak/left"));
              TextureImage left_m = TextureImage.open(manager, EveryCompat.res("model/oak_chest_left_m"));
              TextureImage left_o = TextureImage.open(manager, EveryCompat.res("model/oak_chest_left_o"));
-             TextureImage right = TextureImage.open(manager, modRes("model/chest/oak/right"));
+             TextureImage right = TextureImage.open(manager, modRes("quark_variant_chests/oak/right"));
              TextureImage right_m = TextureImage.open(manager, EveryCompat.res("model/oak_chest_right_m"));
              TextureImage right_o = TextureImage.open(manager, EveryCompat.res("model/oak_chest_right_o"));
              TextureImage left_t = TextureImage.open(manager, EveryCompat.res("model/trapped_chest_left"));